Add rest methods for Car class

+++ src/main/java/com/example/cargoTransportation/models/Place.java
@@ -14,8 +14,8 @@ public class Place {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer id;
     private String address;
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
 
     public Place() {
 
@@ -33,16 +33,20 @@ public class Place {
     public void setAddress(String address) {
         this.address = address;
     }
-    public Float getLatitude() {
+
+    public Double getLatitude() {
         return latitude;
     }
-    public void setLatitude(Float latitude) {
+
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
-    public Float getLongitude() {
+
+    public Double getLongitude() {
         return longitude;
     }
-    public void setLongitude(Float longitude) {
+
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
